+++ src/lib.rs
@@ -427,12 +427,12 @@ pub mod systems {
     pub use hopper_core::prelude_core::*;
     pub use hopper_runtime::CpiAccount;
     pub use hopper_runtime::{
-        default_allocator, fast_entrypoint, hopper_entrypoint, hopper_fast_entrypoint,
-        hopper_lazy_entrypoint, lazy_entrypoint, no_allocator, nostd_panic_handler,
-        program_entrypoint, AccountProof, BoundedString, BoundedVec, ExecutableChecked,
-        ExplainExternal, ExternalBytes, ExternalChecked, ExternalExplainSink, ExternalLens,
-        ExternalLensValue, ExternalProof, ExternalResolve, HasOneChecked, HopperString, HopperVec,
-        InstructionAccount, InstructionView, LayoutChecked, OwnerChecked,
+        default_allocator, fast_entrypoint, hopper_entrypoint, hopper_exact_entrypoint,
+        hopper_fast_entrypoint, hopper_lazy_entrypoint, lazy_entrypoint, no_allocator,
+        nostd_panic_handler, program_entrypoint, AccountProof, BoundedString, BoundedVec,
+        ExecutableChecked, ExplainExternal, ExternalBytes, ExternalChecked, ExternalExplainSink,
+        ExternalLens, ExternalLensValue, ExternalProof, ExternalResolve, HasOneChecked,
+        HopperString, HopperVec, InstructionAccount, InstructionView, LayoutChecked, OwnerChecked,
         RemainingExternalAccounts, RemainingGroup, RemainingLazy, RemainingLazySlot, Seed,
         SeedsChecked, SignerChecked, StoredAccountMeta, StoredInstruction, TailCodec, TailElement,
         TokenExtensionsChecked, Unchecked, WritableChecked,
@@ -807,8 +807,8 @@ macro_rules! hopper_dynamic_fields {
 // `hopper_runtime::`.
 pub use hopper_runtime::{
     address, declare_id, default_allocator, err, error, fast_entrypoint, hopper_emit_cpi,
-    hopper_entrypoint, hopper_fast_entrypoint, hopper_lazy_entrypoint, hopper_log,
-    hopper_unsafe_region, lazy_entrypoint, msg, no_allocator, nostd_panic_handler,
+    hopper_entrypoint, hopper_exact_entrypoint, hopper_fast_entrypoint, hopper_lazy_entrypoint,
+    hopper_log, hopper_unsafe_region, lazy_entrypoint, msg, no_allocator, nostd_panic_handler,
     program_entrypoint, require, require_eq, require_gt, require_gte, require_keys_eq,
     require_keys_neq, require_lt, require_lte, require_neq,
 };